+++ Sources/Support/Helpers/UIButton+BackgroundColor.swift
@@ -1,0 +1,26 @@
+/**
+ *  BulletinBoard
+ *  Copyright (c) 2017 - present Alexis Aubry. Licensed under the MIT license.
+ */
+
+import UIKit
+
+extension UIButton {
+
+    /**
+     * Sets a solid background color for the button.
+     */
+
+    func setBackgroundColor(_ color: UIColor, forState controlState: UIControlState) {
+
+        UIGraphicsBeginImageContext(CGSize(width: 1, height: 1))
+        UIGraphicsGetCurrentContext()?.setFillColor(color.cgColor)
+        UIGraphicsGetCurrentContext()?.fill(CGRect(x: 0, y: 0, width: 1, height: 1))
+        let colorImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        setBackgroundImage(colorImage, for: controlState)
+
+    }
+
+}
+
