+++ Sources/Support/Views/Internal/BulletinBackgroundView.swift
