+++ Sources/Support/Views/Internal/HighlightButton.swift
@@ -69,57 +69,3 @@ class HighlightButton: RoundedButton, HighlighterTarget {
     }
 
 }
-
-extension UIButton {
-
-    /**
-     * Sets a solid background color for the button.
-     */
-
-    func setBackgroundColor(_ color: UIColor, forState controlState: UIControlState) {
-
-        UIGraphicsBeginImageContext(CGSize(width: 1, height: 1))
-        UIGraphicsGetCurrentContext()?.setFillColor(color.cgColor)
-        UIGraphicsGetCurrentContext()?.fill(CGRect(x: 0, y: 0, width: 1, height: 1))
-        let colorImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        setBackgroundImage(colorImage, for: controlState)
-
-    }
-
-}
-
-/**
- * A view that wraps a HighlightButton.
- *
- * A wrapper is required to avoid alpha animation issues when unhighlighting the button and performing
- * a bulletin transition.
- */
-
-@objc public class HighlightButtonWrapper: UIView {
-
-    @objc public let button: UIButton
-
-    public required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) is unavailable. Use init(button:) instead.")
-    }
-
-    init(button: HighlightButton) {
-
-        self.button = button
-        super.init(frame: .zero)
-
-        addSubview(button)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-        button.trailingAnchor.constraint(equalTo: trailingAnchor).isActive = true
-        button.topAnchor.constraint(equalTo: topAnchor).isActive = true
-        button.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
-
-    }
-
-    public override var intrinsicContentSize: CGSize {
-        return button.intrinsicContentSize
-    }
-
-}
