+++ Sources/Support/Views/Internal/Highlighter.swift
@@ -5,13 +5,27 @@
 
 import UIKit
 
+/// An item that can be highlighted due to a touch event.
 protocol HighlighterTarget: class {
+
+    /// Highlight the item.
     func highlight()
+
+    /// Unhighlight the item.
     func unhighlight()
+
 }
+
+/**
+ * An object that determines when an item (its target) needs to be highlighted in response to
+ * touch events.
+ *
+ * You must forward the target's `touchesBegan`, `touchesMoved` and `touchesEnded` to this highlighter.
+ */
 
 class Highlighter {
 
+    /// The control targeted for highlight.
     weak var target: (UIControl & HighlighterTarget)?
 
     // MARK: - Hit Area
@@ -24,20 +38,31 @@ class Highlighter {
         return bounds.applying(scaleTransform).applying(translateTransform)
     }
 
-    /// The area where touches are considered to be inside the button.
     private var hitArea: CGRect {
         return _cachedHitArea ?? makeHitArea(bounds: target!.bounds)
     }
 
     // MARK: - Touch Handling
 
+    /**
+     * Call this method when the bounds of the target change.
+     */
+
     func updateBounds(_ bounds: CGRect) {
         _cachedHitArea = makeHitArea(bounds: bounds)
     }
 
+    /**
+     * Call this method when touches begin on the target.
+     */
+
     func handleTouchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         target?.highlight()
     }
+
+    /**
+     * Call this method when touches move on the target.
+     */
 
     func handleTouchesMoved(_ touches: Set<UITouch>, with event: UIEvent?) {
 
@@ -61,6 +86,10 @@ class Highlighter {
         }
 
     }
+
+    /**
+     * Call this method when touches end on the target.
+     */
 
     func handleTouchesEnded(_ touches: Set<UITouch>, with event: UIEvent?) {
